requirements.txt
arrow
python-dotenv
rich
python-frontmatter
opensearch-py
langchain
sentence-transformers